content-security-policy/navigation/javascript-url-navigation-to-child-snapshots-parent-csp.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <iframe id="iframe"></iframe>
  <script nonce="abc">
    // The intention of this test is to ensure a snapshot of the CSP from the
    // time of creating the `javascript:` URL navigation tasks is used when
    // executing the tasks.
    // The spec is buggy, see
    // <https://github.com/whatwg/html/issues/4651#issuecomment-2412623188>
    // and the encompassing ticket. This is one step towards correcting it.

    setup({ single_test: true });

    let numberOfAllowedCalls = 0;

    function f() {
      assert_true(true, "`f()` is called");
      ++numberOfAllowedCalls;
    }

    function g() {
      assert_unreached("`g()` shouldn't be called");
    }

    const iframe = document.getElementById("iframe");

    function onCSPWithScriptSrcNonceAbcAdded() {
      // Schedule another task, according to step 20 of
      // <https://html.spec.whatwg.org/#navigate>.
      iframe.contentWindow.location.href = "javascript: parent.g()";

      document.addEventListener("securitypolicyviolation", (e) => {
          assert_equals(e.effectiveDirective, "script-src-elem");
          assert_equals(e.blockedURI, "inline");

          // To avoid this test from errorneously passing.
          assert_equals(numberOfAllowedCalls, 2, "`f()` was called");

          done();
      });
    }

    function addCSPWithFrameSrcNonceAbcToDoc() {
      const m = document.createElement("meta");
      m.setAttribute("http-equiv", "content-security-policy");
      m.setAttribute("content", "script-src 'nonce-abc'");
      document.head.append(m);

      onCSPWithScriptSrcNonceAbcAdded();
    }

    window.addEventListener("load", () => {
      // Schedule tasks, according to step 20 of
      // <https://html.spec.whatwg.org/#navigate>.
      iframe.contentWindow.location.href = "javascript:parent.f()";
      iframe.contentWindow.location.href =
        "javascript:parent.addCSPWithFrameSrcNonceAbcToDoc()";
      iframe.contentWindow.location.href = "javascript:parent.f()";
    });

  </script>
</body>
</html>
